fix: hooks rename (2/2)
